// frontend/src/features/actividades/corregir_reparos_gestor/models/corregir_reparos_gestor.ts
export interface CorregirReparosGestor {
  id_corregir_reparos_gestor: number;
  id_expediente: number;
  id_usuario_solicitante: number;
  is_subsanar: boolean;
  observaciones?: string | null;
  estatus_general?: string | null;

  solicitante?: string | null;
  observaciones_reparo?: string | null;
  fecha_ingreso?: string | null;

  is_active: boolean;
  row_status: boolean;
  created_by: number;
  created_date: string;
  modified_by?: number | null;
  modified_date?: string | null;
}